+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.13] - 2023-01-23
+
+### Fixed
+- The Command `circleci-continue-long-job-cancel` did not have the ability to properly paginate
+  over results. This release makes it such that the Cancel command will now walk over a period of
+  time, instead of a set number of pages.
 
 ## [1.0.12] - 2023-01-20
 ### Added